refactor(core): drop @Routine decorator in favour of the Routine class
The `@Routine` decorator was a visual-symmetry-only wrapper around the
class-less `Routine` API. The class body was never used — `@Property`
on a routine class was documented as ignored, there were no field- or
method-as-param semantics, and the entire routine definition lived in
the decorator's config bag. The decorator only contributed `target.name`
(metadata key) and a class reference (identity), both of which the
`Routine` class instance already provides with the same generic carry
for arg/return types.

Keeping both paths invited the footgun of users writing `@Property()`
fields on a routine class and being surprised when nothing happened.
This was unreleased on `feat/stored-routines`, so the cleanup is in
scope of the same feature work — no breaking-change concern.

Removed:
- `packages/decorators/src/{legacy,es}/Routine.ts` and their re-exports
- `getRoutineMetadataFromDecorator` from `decorators/utils.ts`
- `findGlobalRoutineForClass` / class-typeof branch in
  `MetadataDiscovery.collectRoutines`
- Static `MetadataStorage.getRoutineMetadata` registry (only consumer
  was the decorator path)
- Decorator-style emission in `RoutineSourceFile` — it now always emits
  `export const X = new Routine({...})` regardless of `entityDefinition`
- `@Routine` mentions across JSDoc, the stored-routines doc page, and
  the decorators doc page

Tests: deleted the ES-decorator routine test and renamed
`stored-routines-decorator.test.ts` → `stored-routines-metadata.test.ts`
keeping the still-relevant metadata-edge tests; updated the
entity-generator routine tests to assert the class-form output.
