+++ packages/core/src/connections/Connection.ts
@@ -167,9 +167,9 @@ export abstract class Connection {
   ): Promise<QueryResult<T> | any | any[]>;
 
   /**
-   * Invokes a stored procedure or function declared via the `@Routine` decorator or the
-   * {@link Routine} class. The base implementation throws — only drivers that override this
-   * method (every SQL driver does) support routine invocation.
+   * Invokes a stored procedure or function declared via the {@link Routine} class. The base
+   * implementation throws — only drivers that override this method (every SQL driver does)
+   * support routine invocation.
    *
    * @internal - public callers should go through `EntityManager.callRoutine`, which resolves the
    *   metadata, dispatches to the driver, and surfaces OUT/INOUT writes via ScalarReference.
