+++ packages/decorators/src/legacy/index.ts
@@ -7,7 +7,6 @@ export * from './OneToMany.js';
 export * from './Property.js';
 export * from './Check.js';
 export * from './Trigger.js';
-export * from './Routine.js';
 export * from './Enum.js';
 export * from './Formula.js';
 export * from './Indexed.js';
